% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomaRt-package.R, R/biomaRt.R,
%   R/utilityFunctions.R
\name{biomaRt-deprecated}
\alias{biomaRt-deprecated}
\alias{filterOptions}
\alias{searchFilterValues}
\alias{listFilterValues}
\title{Deprecated and defunct functions in package \sQuote{biomaRt}}
\usage{
filterOptions(filter, mart)

searchFilterValues(mart, filter, pattern)

listFilterValues(mart, filter)
}
\description{
These functions have been removed from biomaRt and replaced with
alternatives.
}
\details{
The following functions are defunct and no longer work; use the replacement
indicated below:

\itemize{
\item{filterOptions: \code{\link{listFilterOptions}}}
\item{listFilterValues: \code{\link{listFilterOptions}}}
\item{searchFilterValues: \code{\link{searchFilterOptions}}}
}
}
