% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{biomartCache}
\alias{biomartCache}
\alias{biomartCacheClear}
\alias{biomartCacheInfo}
\title{\pkg{biomaRt} result caching}
\usage{
biomartCacheClear()

biomartCacheInfo()
}
\value{
These functions do not return anything and are called for their side
effects. \code{\link[=biomartCacheInfo]{biomartCacheInfo()}} prints the location of the cache, along
with the number of files and their total size on disk.
\code{\link[=biomartCacheClear]{biomartCacheClear()}} will delete the current contents of the cache.
}
\description{
\pkg{biomaRt} makes use of a results cache to speedup execution of queries
that have been run before. These functions provide details on the status of
this cache, and allow it to be deleted.
}
\author{
Mike Smith
}
\keyword{IO}
