% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensembl_versions}
\alias{ensembl_versions}
\title{Mapping of Ensembl versions to their release dates}
\format{
A \code{data.frame} with two columns:
\itemize{
\item \code{version}: Ensembl version (e.g., "v116")
\item \code{release_date}: Release date of the Ensembl version (e.g., "June 2026")
}
}
\source{
\url{https://jun2026.archive.ensembl.org/info/website/archives/assembly.html}
}
\usage{
data(ensembl_versions)
}
\description{
Mapping of Ensembl versions to their release dates
}
\keyword{datasets}
