\name{GO-class}
\doctype{class}
\alias{GO-class}
\title{Class GO}
\description{Represents a GO annotation class}
\section{Creating Objects}{
\code{ new('GO',}\cr
\code{	GOID = ..../, #GO identfier of class character}\cr
\code{	description = ..../, #gene description of class character}\cr
\code{	evidence = ..../, #evidence of class character}\cr
\code{ )}}
\details{
	This Class is used to represent a GO annotation.  It stores the GO identifier,
	a description of the GO term, the chromosome and evidence fo the annotation.
}
\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 

\seealso{}

\examples{


}
\keyword{methods}

