% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomaRt.R, R/utilityFunctions.R
\name{listDatasets}
\alias{listDatasets}
\alias{searchDatasets}
\title{List or search the datasets available in the selected BioMart database}
\usage{
listDatasets(mart, verbose = FALSE, useCache = TRUE)

searchDatasets(mart, pattern = ".*")
}
\arguments{
\item{mart}{object of class Mart created with the useMart function}

\item{verbose}{Give detailed output of what the method is doing, for
debugging purposes}

\item{useCache}{If \code{useCache = TRUE} (the default) biomaRt will try to store
successful query results on disk, and will load these if a query is run
again, rather than contacting the upstream server.}

\item{pattern}{Character vector defining the regular expression
(\link[base:regex]{regex}) to be used for the search.  If left blank the
default is to use ".*" which will match everything and return the same as
\code{\link[=listDatasets]{listDatasets()}}.}
}
\description{
Lists or search the datasets available in the selected BioMart database
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
## list the available Ensembl marts and use Ensembl Genes
listEnsembl()
ensembl <- useEnsembl(biomart = "ENSEMBL_MART_ENSEMBL")

## list the available datasets in this Mart
listDatasets(mart = ensembl)

## the list of Ensembl datasets grows ever larger (101 as of Ensembl 93)
## we can search for a term of interest to reduce the length e.g. 'sapiens'
searchDatasets(mart = ensembl, pattern = "sapiens")

## search for any dataset containing the word Rat or rat
searchDatasets(mart = ensembl, pattern = "(R|r)at")
\dontshow{\}) # examplesIf}
}
\author{
Steffen Durinck, Mike Smith
}
\keyword{methods}
