\name{martConnect}
\alias{martConnect}
\title{Connects to available BioMarts}
\description{This function connects to available BioMarts}
\usage{martConnect()}

\arguments{
}

\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 

\seealso{}

\examples{

mart <- martConnect()

martDisconnect(mart = mart)
}
\keyword{methods}

