\name{martTable-class}
\doctype{class}
\alias{martTable-class}
\title{Class martTable}
\description{Represents a martTable class}
\section{Creating Objects}{
\code{ new('martTable',}\cr
\code{	id = ..../, #OMIM identifier of class numeric}\cr
\code{	table = ..../, #list}\cr
\code{ )}}
\details{
	This Class is used to represent an annotation table.  It stores the feature identifier (affy, locuslink,...),
	and a table with annotation retrieved for these features.
}
\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 

\seealso{}

\examples{


}
\keyword{methods}

