\name{OMIM-class}
\doctype{class}
\alias{OMIM-class}
\title{Class OMIM}
\description{Represents a OMIM class}
\section{Creating Objects}{
\code{ new('OMIM',}\cr
\code{	OMIMID = ..../, #OMIM identifier of class numeric}\cr
\code{	disease = ..../, #disease of class character}\cr
\code{ )}}
\details{
	This Class is used to represent OMIM annotation.  It stores the OMIM identifier,
	and the corresponding disease.
}
\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 

\seealso{}

\examples{


}
\keyword{methods}

