\name{parseAttributes}
\alias{parseAttributes}
\title{parse attributes}
\description{Function to parse attributes from BioMart configuration files.  This function is not to be used by BioMart users directly}
\usage{parseAttributes(xml, env, group = "", page = "")}

\arguments{
\item{xml}{xml configuration}
\item{env}{Environment to put the attribute info in}
\item{group}{Pass the attribute group}
\item{page}{Pass the attribute page}
}

\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 
\keyword{methods}

