\name{parseFilters}
\alias{parseFilters}
\title{parse filters}
\description{Function to parse filters from BioMart configuration files.  This function is not to be used by BioMart users directly}
\usage{parseFilters(xml, env)}

\arguments{
\item{xml}{xml configuration}
\item{env}{Environment to put the filter info in}
}
\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 

\seealso{}

\examples{
}
\keyword{methods}

