% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomaRt-package.R
\name{shared_arguments}
\alias{shared_arguments}
\title{Shared arguments for biomaRt functions}
\arguments{
\item{useCache}{If \code{useCache = TRUE} (the default) biomaRt will try to store
successful query results on disk, and will load these if a query is run
again, rather than contacting the upstream server.}
}
\description{
Shared arguments for biomaRt functions
}
\keyword{internal}
