\name{biomaRt-deprecated}
\alias{biomaRt-deprecated}
\alias{filterOptions}
\alias{listFilterValues}
\alias{searchFilterValues}
\alias{getBMlist}
\title{Deprecated and defunct functions in package \sQuote{biomaRt}}
\description{
  These functions have been removed from biomaRt and replaced with alternatives.
}

\details{

  The following functions are defunct and no longer work; use
  the replacement indicated below:
  \itemize{

    \item{filterOptions: \code{\link{listFilterOptions}}}
    \item{listFilterValues: \code{\link{listFilterOptions}}}
    \item{searchFilterValues: \code{\link{searchFilterOptions}}}

  }

}
