% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomaRt-package.R
\name{biomaRt-deprecated}
\alias{biomaRt-deprecated}
\alias{filterOptions}
\alias{searchFilterValues}
\alias{listFilterValues}
\title{Deprecated and defunct functions in package \pkg{biomaRt}}
\description{
These functions have been removed from biomaRt and replaced with
alternatives.
}
\details{
The following functions are defunct and no longer work; use the replacement
indicated below:
\itemize{
\item filterOptions: \code{\link[=listFilterOptions]{listFilterOptions()}}
\item listFilterValues: \code{\link[=listFilterOptions]{listFilterOptions()}}
\item searchFilterValues: \code{\link[=searchFilterOptions]{searchFilterOptions()}}
}
}
