\name{biomaRt-deprecated}
\alias{biomaRt-deprecated}
\alias{filterOptions}
\alias{listFilterValues}
\alias{searchFilterValues}
\alias{getBMlist}
\title{Deprecated functions in package \sQuote{biomaRt}}
\description{
  These functions are provided for compatibility with older versions
  of \sQuote{biomaRt} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{filterOptions: \code{\link{listFilterOptions}}}
    \item{listFilterValues: \code{\link{listFilterOptions}}}
    \item{searchFilterValues: \code{\link{searchFilterOptions}}}

  }
  
  The following functions are defunct and no longer work; use
  the replacement indicated below:
  \itemize{
      \item{getBMlist: \code{\link{getBM}}}
  }
}
