% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomaRt.R
\name{exportFASTA}
\alias{exportFASTA}
\title{Exports getSequence results to FASTA format}
\usage{
exportFASTA(sequences, file)
}
\arguments{
\item{sequences}{A data.frame that was the output of the getSequence
function}

\item{file}{File to which you want to write the data}
}
\description{
Exports getSequence results to FASTA format
}
\examples{


if(interactive()){
    mart <- useMart("ensembl", dataset="hsapiens_gene_ensembl")

    #seq<-getSequence(chromosome=c(2,2),start=c(100000,30000),end=c(100300,30500),mart=mart)
    #exportFASTA(seq,file="test.fasta")

}

}
\author{
Steffen Durinck
}
\keyword{methods}
