% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomaRt.R
\name{filterType}
\alias{filterType}
\title{Displays the filter type}
\usage{
filterType(filter, mart)
}
\arguments{
\item{filter}{A valid filter name. Valid filters are given by the
\code{\link[=listFilters]{listFilters()}} function}

\item{mart}{object of class Mart, created using the \code{\link[=useMart]{useMart()}} function}
}
\description{
Displays the type of the filer given a filter name.
}
\examples{


if(interactive()){
mart = useMart("ensembl", dataset="hsapiens_gene_ensembl")
filterType("chromosome_name", mart)
}

}
\author{
Steffen Durinck
}
\keyword{methods}
