\name{getAffyArrays}
\alias{getAffyArrays}
\title{Displays affy identifiers of Affymetrix chips supported by the Ensembl package}
\description{Displays affy identifiers of Affymetrix chips supported by the Ensembl package}
\usage{getAffyArrays()}

\arguments{
}

\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 

\seealso{}

\examples{
getAffyArrays()

}
\keyword{methods}

