% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{listEnsemblArchives}
\alias{listEnsemblArchives}
\title{Lists the available archived versions of Ensembl}
\usage{
listEnsemblArchives()
}
\description{
Returns a table containing the available archived versions of Ensembl, along
with the dates they were created and the URL used to access them.
}
\examples{

listEnsemblArchives()

}
\author{
Mike Smith
}
\keyword{methods}
