\name{listEnsemblArchives}
\alias{listEnsemblArchives}
\title{Lists the available archived versions of Ensembl}
\description{Returns a table containing the available archived versions of 
Ensembl, along with the dates they were created and the URL used to access
them.}

\usage{listEnsemblArchives()}

\author{Mike Smith} 

\examples{
listEnsemblArchives()
}
\keyword{methods}

