% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Mart.R
\name{show,Mart-method}
\alias{show,Mart-method}
\alias{Mart-class}
\title{Class Mart}
\usage{
\S4method{show}{Mart}(object)
}
\description{
Represents a Mart class, containing connections to different BioMarts
}
\section{Methods}{
 \describe{\code{show} Print summary of the object}
}

\author{
Steffen Durinck
}
\keyword{methods}
