\name{Mart-class}
\doctype{class}
\alias{Mart-class}
\title{Class Mart}
\description{Represents a Mart class, containing connections to different BioMarts}
\section{Creating Objects}{
\code{ new('Mart',}\cr
\code{	connections = ..../, #connections BioMart databases}\cr
\code{	arrayToSpecies = ..../, #mapping from affy arrays to species}\cr
\code{ )}}
\details{
	This Class is used to represent connections to different BioMart databases.  
}
\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 

\seealso{}

\examples{


}
\keyword{methods}

