\name{Mart-class}
\doctype{class}
\alias{Mart-class}
\title{Class Mart}
\description{Represents a Mart connection}
\section{Creating Objects}{
\code{ new('Mart',}\cr
\code{	connection = ..../, #OMIM identifier of class numeric}\cr
\code{	name = ..../, #disease of class character}\cr
\code{ )}}
\details{
	This Class is used to represent a Mart connection.  With a slot containing the connection and slot mart containing the mart name.
}
\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 

\seealso{}

\examples{


}
\keyword{methods}

