\name{Mart-class}
\alias{Mart-class}
\alias{show,Mart-method}
\title{Class Mart}
\description{Represents a Mart class, containing connections to different BioMarts}
\section{Methods}{
   \describe{\code{show} Print summary of the object}
}
\author{Steffen Durinck, \url{http://www.stat.berkeley.edu/~steffen}} 
\keyword{methods}

