% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP2009code.R
\name{NP2009code}
\alias{NP2009code}
\title{Display the analysis code from the 2009 Nature protocols paper}
\usage{
NP2009code()
}
\description{
This function opens an editor displaying the analysis code of the Nature
Protocols 2009 paper
}
\details{
The \code{\link[=edit]{edit()}} function uses \code{getOption("editor")} to select
the editor. Use, for instance, \code{options(editor="emacs")} to set another
editor.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
NP2009code()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=edit]{edit()}}
}
\author{
Steffen Durinck, Wolfgang Huber
}
\keyword{methods}
