\name{NP2009code}
\alias{NP2009code}
\title{Display the analysis code from the 2009 Nature protocols paper}
\description{This function opens an editor displaying the analysis code of the Nature Protocols 2009 paper}
\usage{NP2009code()}

\arguments{
}

\author{Steffen Durinck, \url{http://www.stat.berkeley.edu/~steffen}, Wolfgang Huber} 

\seealso{}

\examples{
if(interactive()){
NP2009code()
}
}
\keyword{methods}

