\name{parseAttributes}
\alias{parseAttributes}
\title{parse attributes}
\description{Function to parse attributes from BioMart configuration files.  This function is not to be used by BioMart users directly}
\usage{parseAttributes(xml, env)}

\arguments{
\item{xml}{xml configuration}
\item{env}{Environment to put the attribute info in}
 

}

\author{Steffen Durinck, \url{http://www.esat.kuleuven.ac.be/~sdurinck}} 

\seealso{}

\examples{
}
\keyword{methods}

