\name{parseFilters}
\alias{parseFilters}
\title{parse filters}
\description{Function to parse filters from BioMart configuration files.  This function is not to be used by BioMart users directly}
\usage{parseFilters(xml, env, group = "", page = "")}

\arguments{
\item{xml}{xml configuration}
\item{env}{Environment to put the filter info in}
\item{group}{Pass the filter group}
\item{page}{Pass the filter page}
}
\author{Steffen Durinck, \url{http://www.stat.berkeley.edu/~steffen}} 
\keyword{methods}

